+++ mobile_app/src/constants.ts
@@ -1,0 +1,4 @@
+// In mobile_app/src/constants.ts
+
+// TODO: Make sure this is still your computer's IP address
+export const API_BASE_URL = 'http://192.168.1.178:8000/api';
